pyproject.toml
[tool.poetry]
name = "pokedex-back"
version = "0.1.0"
description = "Implémentation du back du test technique POKEDEX"
authors = ["Marc PASTOR-ABAD <mpastor@simco.fr>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.0"
djangorestframework = "^3.13.1"
django-filter = "^21.1"
psycopg2-binary = "^2.9.3"
django-env-overrides = "^0.1.0"
django-extensions = "^3.1.5"
django-cors-headers = "^3.11.0"
uwsgi = "^2.0.18"
whitenoise = "^6.0.0"
djangorestframework-simplejwt = "^5.1.0"
drf-spectacular = "^0.22.0"

[tool.poetry.dev-dependencies]
factory-boy = "^3.2.1"
Faker = "^13.3.0"
pre-commit = "^2.17.0"
black = "^22.1.0"
pytest = "^7.0.1"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.1.0"
pytest-xdist = "^2.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
